backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
pydantic
python-multipart
python-dotenv
alembic
python-dotenv
